[metrics] move EnrichTags into tagger
This code is shared between metrics, events and service checks.
