+++ pkg/metrics/metric_sample.go
@@ -6,12 +6,8 @@
 package metrics
 
 import (
-	"github.com/DataDog/datadog-agent/pkg/dogstatsd/packets"
 	"github.com/DataDog/datadog-agent/pkg/tagger"
-	"github.com/DataDog/datadog-agent/pkg/tagger/collectors"
-	"github.com/DataDog/datadog-agent/pkg/telemetry"
 	"github.com/DataDog/datadog-agent/pkg/util"
-	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
 // MetricType is the representation of an aggregator metric type
@@ -35,13 +31,6 @@ var (
 	DistributionMetricTypes = map[MetricType]struct{}{
 		DistributionType: {},
 	}
-
-	// we use to pull tagger metrics in dogstatsd. Pulling it later in the
-	// pipeline improve memory allocation. We kept the old name to be
-	// backward compatible and because origin detection only affect
-	// dogstatsd metrics.
-	tlmUDPOriginDetectionError = telemetry.NewCounter("dogstatsd", "udp_origin_detection_error",
-		nil, "Dogstatsd UDP origin detection error count")
 )
 
 // String returns a string representation of MetricType
@@ -105,45 +94,10 @@ func (m *MetricSample) GetHost() string {
 	return m.Host
 }
 
-func findOriginTags(origin string, cardinality collectors.TagCardinality, tb *util.TagsBuilder) {
-	if origin != packets.NoOrigin {
-		if err := tagger.TagBuilder(origin, cardinality, tb); err != nil {
-			log.Errorf(err.Error())
-		}
-	}
-}
-
-func addOrchestratorTags(cardinality collectors.TagCardinality, tb *util.TagsBuilder) {
-	// Include orchestrator scope tags if the cardinality is set to orchestrator
-	if cardinality == collectors.OrchestratorCardinality {
-		if err := tagger.OrchestratorScopeTagBuilder(tb); err != nil {
-			log.Error(err.Error())
-		}
-	}
-}
-
-// EnrichTags expend a tag list with origin detection tags
-// NOTE(remy): it is not needed to sort/dedup the tags anymore since after the
-// enrichment, the metric and its tags is sent to the context key generator, which
-// is taking care of deduping the tags while generating the context key.
-func EnrichTags(tb *util.TagsBuilder, originID string, k8sOriginID string, cardinality string) {
-	taggerCard := taggerCardinality(cardinality)
-
-	findOriginTags(originID, taggerCard, tb)
-	addOrchestratorTags(taggerCard, tb)
-
-	if k8sOriginID != "" {
-		if err := tagger.TagBuilder(k8sOriginID, taggerCard, tb); err != nil {
-			tlmUDPOriginDetectionError.Inc()
-			log.Tracef("Cannot get tags for entity %s: %s", k8sOriginID, err)
-		}
-	}
-}
-
 // GetTags returns the metric sample tags
 func (m *MetricSample) GetTags(tb *util.TagsBuilder) {
 	tb.Append(m.Tags...)
-	EnrichTags(tb, m.OriginID, m.K8sOriginID, m.Cardinality)
+	tagger.EnrichTags(tb, m.OriginID, m.K8sOriginID, m.Cardinality)
 }
 
 // Copy returns a deep copy of the m MetricSample
@@ -155,18 +109,3 @@ func (m *MetricSample) Copy() *MetricSample {
 	return dst
 }
 
-// taggerCardinality converts tagger cardinality string to collectors.TagCardinality
-// It defaults to DogstatsdCardinality if the string is empty or unknown
-func taggerCardinality(cardinality string) collectors.TagCardinality {
-	if cardinality == "" {
-		return tagger.DogstatsdCardinality
-	}
-
-	taggerCardinality, err := collectors.StringToTagCardinality(cardinality)
-	if err != nil {
-		log.Tracef("Couldn't convert cardinality tag: %w", err)
-		return tagger.DogstatsdCardinality
-	}
-
-	return taggerCardinality
-}
