+++ pkg/metrics/metric_sample_test.go
@@ -9,11 +9,6 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/DataDog/datadog-agent/pkg/tagger"
-	"github.com/DataDog/datadog-agent/pkg/tagger/collectors"
-	"github.com/DataDog/datadog-agent/pkg/tagger/local"
-	"github.com/DataDog/datadog-agent/pkg/util"
-
 	"github.com/stretchr/testify/assert"
 )
 
@@ -33,59 +28,3 @@ func TestMetricSampleCopy(t *testing.T) {
 	assert.True(t, reflect.DeepEqual(&src, &dst))
 }
 
-func Test_taggerCardinality(t *testing.T) {
-	tests := []struct {
-		name        string
-		cardinality string
-		want        collectors.TagCardinality
-	}{
-		{
-			name:        "high",
-			cardinality: "high",
-			want:        collectors.HighCardinality,
-		},
-		{
-			name:        "orchestrator",
-			cardinality: "orchestrator",
-			want:        collectors.OrchestratorCardinality,
-		},
-		{
-			name:        "orch",
-			cardinality: "orch",
-			want:        collectors.OrchestratorCardinality,
-		},
-		{
-			name:        "low",
-			cardinality: "low",
-			want:        collectors.LowCardinality,
-		},
-		{
-			name:        "empty",
-			cardinality: "",
-			want:        tagger.DogstatsdCardinality,
-		},
-		{
-			name:        "unknown",
-			cardinality: "foo",
-			want:        tagger.DogstatsdCardinality,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.want, taggerCardinality(tt.cardinality))
-		})
-	}
-}
-
-func TestEnrichTagsOrchestrator(t *testing.T) {
-	oldTagger := tagger.GetDefaultTagger()
-	defer tagger.SetDefaultTagger(oldTagger)
-
-	fakeTagger := local.NewFakeTagger()
-	tagger.SetDefaultTagger(fakeTagger)
-	fakeTagger.SetTags("foo", "fooSource", []string{"lowTag"}, []string{"orchTag"}, nil, nil)
-
-	tb := util.NewTagsBuilder()
-	EnrichTags(tb, "foo", "", "orchestrator")
-	assert.Equal(t, []string{"lowTag", "orchTag"}, tb.Get())
-}
